Split out loginSocialUser
